Refactor dataset.py to add encoder and decoder masks
The code changes in dataset.py add encoder and decoder masks to the dataset class. This change includes the addition of "encoder_mask" and "decoder_mask" keys to the returned dictionary in the __getitem__ method. The encoder mask is computed by checking if the encoder input is not equal to the pad token, while the decoder mask is computed using the causal_mask function and checking if the decoder input is not equal to the pad token. These masks are important for sequence-to-sequence models and will be used during training and inference.